under construction!
doing a custom design is half done and my first post is still going to be a while,
so putting up an under construction page instead. also getting rid of the old "first"
article, because it was a bit too off the cuff and goes right against what new "first"
post is going to be about.